refactor: extract DataTableRows class for better organization and reusability
