+++ test/OpenFeature.E2ETests/Steps/MetadataStepDefinitions.cs
@@ -66,19 +66,4 @@ public class MetadataStepDefinitions : BaseStepDefinitions
                 throw new ArgumentOutOfRangeException();
         }
     }
-
-    private class DataTableRows
-    {
-        public DataTableRows(string key, string value, string metadataType)
-        {
-            this.Key = key;
-            this.Value = value;
-
-            this.MetadataType = FlagTypesUtil.ToEnum(metadataType);
-        }
-
-        public string Key { get; }
-        public string Value { get; }
-        public FlagType MetadataType { get; }
-    }
 }
